refactor: move to screen
