+++ src/i18n.ts
@@ -36,7 +36,7 @@ export enum I18N {
   transactionAddressPlaceholder,
   transactionAddressError,
   transactionAddressLabel,
-  TransactionEmpty,
+  transactionsEmpty,
   trackActivityTitle,
   trackActivityImprovement,
   trackActivityImprovementDescription,
@@ -129,7 +129,7 @@ export enum I18N {
   settingsTestStaked,
   settingsTestUnbounded,
   settingsTestUndelegate,
-  sent,
+  transactionSendTitle,
   settingsContactEditDelete,
   settingsContactEditSure,
   stakingValidators,
@@ -222,7 +222,7 @@ export enum I18N {
   transactionSumSend,
   transactionSumPereview,
   transactions,
-  to,
+  transactionSendTo,
   settingsAccountNoWallet,
   settingsAddressBookLabel,
   settingsAddressBookPlaceholder,
@@ -319,7 +319,7 @@ const en: Record<I18N, string> = {
   [I18N.transactionAddressPlaceholder]: 'Address (0x) or contact name',
   [I18N.transactionAddressError]: 'Incorrect address',
   [I18N.transactionAddressLabel]: 'Send to',
-  [I18N.TransactionEmpty]: 'No transactions',
+  [I18N.transactionsEmpty]: 'No transactions',
   [I18N.trackActivityTitle]: 'Allow ISLM Wallet track your activity',
   [I18N.trackActivityImprovement]: 'Improvement',
   [I18N.trackActivityImprovementDescription]:
@@ -434,7 +434,7 @@ const en: Record<I18N, string> = {
   [I18N.SettingsSecurityChangePin]: 'Change PIN',
   [I18N.setttingsSecurityEnterPin]: 'Enter new PIN',
   [I18N.settingsSecurityWalletPin]: 'Enter ISLM Wallet PIN',
-  [I18N.sent]: 'Sent',
+  [I18N.transactionSendTitle]: 'Sent',
   [I18N.stakingValidators]: 'Validators list',
   [I18N.stakingValidatorsRowPower]: 'Power: {{power}}',
   [I18N.stakingValidatorsRowStaked]: 'Staked: {{staked}}',
@@ -532,7 +532,7 @@ const en: Record<I18N, string> = {
   [I18N.transactionSumSend]: 'Send to',
   [I18N.transactionSumPereview]: 'Preview',
   [I18N.transactions]: 'Transactions',
-  [I18N.to]: 'to {{address}}',
+  [I18N.transactionSendTo]: 'to {{address}}',
   [I18N.settingsAccountNoWallet]: 'No wallets',
   [I18N.settingsAddressBookLabel]: 'Address',
   [I18N.settingsAddressBookPlaceholder]: 'Search or add a contact',
